requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth
fastapi
uvicorn[standard]